// src/lib/types/enums.ts
export enum Methods {
  AGGREGATE = 'aggregate',
  UPDATE = 'update',
  UPDATE_MANY = 'updateMany',
  INSERT = 'insert',
  INSERT_MANY = 'insertMany',
  FIND_MANY = 'findMany',
  DELETE_MANY = 'deleteMany',
  BULK_WRITE = 'bulkWrite',
  FIND = 'find',
  FIND_BY_ID = 'findById',
  DELETE = 'delete',
  TOTAL = 'total',
}

export enum CUSTOM_VALIDATION {
  UNIQUE = 'unique',
}